getSortBy & getPageable methods should be reusable for public pages. Hence moving them to BaseResource
